⚡️ Speed up method `PlayQueueService.clear_queue` by 3,264%

+++ backend/services/playqueue_service.py
@@ -60,5 +60,6 @@ class PlayQueueService:
 
     @staticmethod
     def clear_queue() -> PlayQueue:
-        db.truncate()
+        if db.all():
+            db.truncate()
         return PlayQueue()
